Rework command log data modeling
This is a change to how we plan to represent game input.
Previously, a Command struct contained all data necessary to begin
execution from that point in the game. This was unwieldy, as we had
repeated data: CardPlayCommand had player+card, which made an
OpsCommand and passed down player+card, etc.

This went against the idea of only storing variable data in the command
log, and relying on global state for everything else.

Instead of embedding all local state in the Command (now Input) struct,
we will use normal local variables and functions to execute the game,
making Input structs only to capture data that must be written to the
save log.

    * rename 'Command' to 'Input' throughout
    * Get rid of testing-oriented code in main.go
    * gofmt some things we missed
    * rename Card/Country .Name() to String()
    * support un/marshaling PlayKind, OpsKind
    * fix lookupAff to always expect lowercase
    * add func NewState