UI.Message, remove Input interface, etc
The original goal of interface Input was to transparently support
requesting new input or replaying old input from a save file, but
the old implementation didn't actually accomplish that. E.g. what
if the "input" was a die roll? That is something that must be
saved in the log for later replay, but doesn't require actual user
input.
Remove SelectCountry and SelectCard in ui.go
State embeds UI
TerminalUI embeds a bytes.Buffer